requirements.txt
streamlit
pdfplumber
pytesseract
pillow
pandas
requests
